requirements.txt
# AI/ML Core Libraries
tensorflow>=2.13.0
keras>=2.13.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
scikit-learn>=1.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0

# NLP & Conversational AI
rasa>=3.6.0
transformers>=4.30.0
sentencepiece>=0.1.99

# Indic Language Support
indic-nlp-library>=0.81
indic-transliteration>=2.3.0

# Speech Processing
openai-whisper>=20230314
coqui-tts>=0.13.0
soundfile>=0.12.0

# Explainable AI
lime>=0.2.0.1
grad-cam>=1.4.0

# Web Interface
streamlit>=1.25.0
streamlit-chat>=0.1.1
streamlit-option-menu>=0.3.6
streamlit-extras>=0.3.0
plotly>=5.15.0
altair>=5.0.0

# API Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# PostgreSQL Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
bcrypt>=4.0.0

# Email Validation
email-validator>=2.0.0

# Google Gemini AI
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
tqdm>=4.65.0

# Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-xdist>=3.3.0
coverage>=7.3.0
locust>=2.17.0
selenium>=4.15.0
pytest-html>=4.1.0
pytest-benchmark>=4.0.0

